+++ src/terrapin/Line.java
@@ -29,13 +29,13 @@ import processing.core.PApplet;
  */
 public class Line {
 	/** From location in x axis. */
-	public int fromX;
+	public float fromX;
 	/** From location in y axis. */
-	public int fromY;
+	public float fromY;
 	/** To location in x axis. */
-	public int toX;
+	public float toX;
 	/** To location in y axis. */
-	public int toY;
+	public float toY;
 	/**
 	 * Colour to draw line in, as a colour created with
 	 * {@link processing.core.PApplet#color(int, int, int)}.
@@ -58,7 +58,7 @@ public class Line {
 	 * @param toY
 	 *            to location in y axis.
 	 */
-	public Line(PApplet applet, int fromX, int fromY, int toX, int toY, int drawColor) {
+	public Line(PApplet applet, float fromX, float fromY, float toX, float toY, int drawColor) {
 		this.applet = applet;
 		this.fromX = fromX;
 		this.fromY = fromY;
